+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
   - upgrade to geantyref 1.3.14
   - removes the core dependency on the caffeine library. This now uses a simple LRU cache for sql parser and sql statements.
+  - adds pluggable cache implementation using caffeine. The old caching behavior can now be restored by using the
+    `jdbi3-caffeine-cache` dependency and adding `jdbi.installPlugin(new CaffeineCachePlugin());`.
 
 # 3.36.0