+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
   - adds pluggable cache implementation using caffeine. The old caching behavior can now be restored by using the
     `jdbi3-caffeine-cache` dependency and adding `jdbi.installPlugin(new CaffeineCachePlugin());`.
   - adds pluggable no-op cache implementation for testing and debugging
+  - improve PostGIS null value handling
 
 # 3.36.0